+++ src/main/java/com/buuz135/industrial/IndustrialForegoing.java
@@ -2,11 +2,8 @@ package com.buuz135.industrial;
 
 import com.buuz135.industrial.proxy.BlockRegistry;
 import com.buuz135.industrial.proxy.CommonProxy;
-import com.buuz135.industrial.proxy.FluidsRegistry;
 import com.buuz135.industrial.utils.IFFakePlayer;
 import com.buuz135.industrial.utils.Reference;
-import com.buuz135.industrial.utils.StrawUtils;
-import com.buuz135.industrial.utils.drinkhandlers.*;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
@@ -24,7 +21,7 @@ import net.ndrei.teslacorelib.items.gears.CoreGearType;
 import java.util.Arrays;
 import java.util.HashMap;
 
-@Mod(modid = Reference.MOD_ID, name = Reference.MOD_ID, version = Reference.VERSION, dependencies = "required-after:teslacorelib", guiFactory = Reference.GUI_FACTORY)
+@Mod(modid = Reference.MOD_ID, name = Reference.MOD_ID, version = Reference.VERSION, dependencies = "required-after:teslacorelib;before:crafttweaker", guiFactory = Reference.GUI_FACTORY)
 public class IndustrialForegoing {
 
     public static CreativeTabs creativeTab = new CreativeTabs(Reference.MOD_ID) {
@@ -52,17 +49,6 @@ public class IndustrialForegoing {
         return null;
     }
 
-    public static void registerDrinkHandlers() {
-        StrawUtils.register(FluidRegistry.WATER, new DrinkHandlerWater());
-        StrawUtils.register(FluidRegistry.LAVA, new DrinkHandlerLava());
-        StrawUtils.register(FluidsRegistry.BIOFUEL, new DrinkHandlerBiofuel());
-        StrawUtils.register(FluidsRegistry.SLUDGE, new DrinkHandlerSludge());
-        StrawUtils.register(FluidsRegistry.SEWAGE, new DrinkHandlerSewage());
-        StrawUtils.register(FluidsRegistry.MILK, new DrinkHandlerMilk());
-        StrawUtils.register(FluidsRegistry.ESSENCE, new DrinkHandlerEssence());
-        StrawUtils.register(FluidsRegistry.MEAT, new DrinkHandlerMeat());
-    }
-
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
         proxy.preInit(event);
@@ -75,7 +61,6 @@ public class IndustrialForegoing {
 
     @Mod.EventHandler
     public void postInit(FMLPostInitializationEvent event) {
-        registerDrinkHandlers();
         proxy.postInit();
     }