+++ src/main/java/com/buuz135/industrial/api/IndustrialForegoingHelper.java
@@ -4,6 +4,7 @@ package com.buuz135.industrial.api;
 import com.buuz135.industrial.api.plant.IPlantRecollectable;
 import com.buuz135.industrial.api.recipe.BioReactorEntry;
 import com.buuz135.industrial.api.recipe.LaserDrillEntry;
+import com.buuz135.industrial.api.recipe.SludgeEntry;
 import net.minecraft.item.ItemStack;
 
 import java.util.stream.Collectors;
@@ -41,6 +42,22 @@ public class IndustrialForegoingHelper {
     public static boolean removeLaserDrillEntry(ItemStack stack) {
         if (LaserDrillEntry.LASER_DRILL_ENTRIES.stream().anyMatch(entry -> stack.isItemEqual(entry.getStack()))) {
             LaserDrillEntry.LASER_DRILL_ENTRIES = LaserDrillEntry.LASER_DRILL_ENTRIES.stream().filter(entry -> !entry.getStack().isItemEqual(stack)).collect(Collectors.toList());
+        }
+        return false;
+    }
+
+    public static boolean addSludgeRefinerEntry(SludgeEntry entry) {
+        if (SludgeEntry.SLUDGE_RECIPES.stream().noneMatch(entry1 -> entry1.getStack().isItemEqual(entry.getStack()))) {
+            SludgeEntry.SLUDGE_RECIPES.add(entry);
+            return true;
+        }
+        return false;
+    }
+
+    public static boolean removeSludgeRefinerEntry(ItemStack stack) {
+        if (SludgeEntry.SLUDGE_RECIPES.stream().anyMatch(entry -> entry.getStack().isItemEqual(stack))) {
+            SludgeEntry.SLUDGE_RECIPES = SludgeEntry.SLUDGE_RECIPES.stream().filter(entry -> !entry.getStack().isItemEqual(stack)).collect(Collectors.toList());
+            return true;
         }
         return false;
     }