+++ src/main/java/com/buuz135/industrial/jei/JEICustomPlugin.java
@@ -54,9 +54,9 @@ public class JEICustomPlugin implements IModPlugin {
 
     @Override
     public void register(IModRegistry registry) {
-        int maxWeight = WeightedRandom.getTotalWeight(BlockRegistry.sludgeRefinerBlock.getItemStackWeightedItems());
+        int maxWeight = WeightedRandom.getTotalWeight(BlockRegistry.sludgeRefinerBlock.getItems());
         List<SludgeRefinerRecipeWrapper> wrapperList = new ArrayList<>();
-        BlockRegistry.sludgeRefinerBlock.getItemStackWeightedItems().forEach(itemStackWeightedItem -> wrapperList.add(new SludgeRefinerRecipeWrapper(itemStackWeightedItem, maxWeight)));
+        BlockRegistry.sludgeRefinerBlock.getItems().forEach(itemStackWeightedItem -> wrapperList.add(new SludgeRefinerRecipeWrapper(itemStackWeightedItem, maxWeight)));
         registry.addRecipes(wrapperList, sludgeRefinerRecipeCategory.getUid());
         registry.addRecipeCatalyst(new ItemStack(BlockRegistry.sludgeRefinerBlock), sludgeRefinerRecipeCategory.getUid());