+++ src/main/java/com/buuz135/industrial/utils/apihandlers/crafttweaker/CTAction.java
@@ -1,0 +1,5 @@
+package com.buuz135.industrial.utils.apihandlers.crafttweaker;
+
+public enum CTAction {
+    ADD, REMOVE;
+}